backend/requirements.txt
asgiref==3.6.0
Django==3.2.18
django-cors-headers==3.14.0
djangorestframework==3.12.4
djoser
django-filter
gunicorn==20.0.4
Pillow==8.3.1
pytest-django
pytest-pythonpath
psycopg2-binary==2.8.6
python-dotenv~=0.21.1
reportlab
